refactor(data-pipeline): preserve existing module layout
