+++ libdd-data-pipeline/src/agentless/config.rs
@@ -1,0 +1,6 @@
+// Copyright 2024-Present Datadog, Inc. https://www.datadoghq.com/
+// SPDX-License-Identifier: Apache-2.0
+
+//! Agentless APM trace export configuration.
+
+pub use libdd_data_pipeline_core::{AgentlessTraceConfig, DEFAULT_AGENTLESS_TIMEOUT};
