+++ libdd-data-pipeline/src/agentless/mod.rs
@@ -23,9 +23,7 @@
 //!   any of them together causes `build`/`build_async` to return
 //!   `BuilderErrorKind::InvalidConfiguration`.
 
+pub(crate) mod config;
 pub(crate) mod exporter;
 
-pub use exporter::send_agentless_traces_http;
-pub use libdd_data_pipeline_core::{
-    AgentlessTraceConfig, TracerMetadata, DEFAULT_AGENTLESS_TIMEOUT,
-};
+pub use config::AgentlessTraceConfig;
